+++ app/models/option_configs/external_identifier_options.rb
@@ -1,6 +1,11 @@
 module OptionConfigs
   class ExternalIdentifierOptions < ExtraOptions
-    def self.raise_bad_configs option_configs
+    def self.set_defaults(config_obj, all_options = {})
+      all_options[:default] ||= {}
+      all_options[:default][:fields] ||= config_obj.all_implementation_fields
+    end
+
+    def self.raise_bad_configs(option_configs)
       # None defined - override with real checks
       # @todo
     end